% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipses.R
\name{ellipses}
\alias{ellipses}
\title{Create general elipse perimeter coordinates}
\usage{
ellipses(x, y, angle, length, width, res = 50)
}
\arguments{
\item{x}{Elipse centre x coordinate}

\item{y}{Elipse centre y coordinate}

\item{angle}{Angle (degrees) of rotation (measured from positive x to the elipse major axis)}

\item{length}{Length of major axis}

\item{width}{Length of minor axis}

\item{res}{Number of points in elipse perimeter}
}
\value{
A dataframe
}
\description{
Parameterised elipse equations. Provide an origin (x and y), minor and major radii,
an angle of rotation (measured from positive x to the elipse major axis) and a resolution.
The funtion returns a list of the x and y coordinates of the perimeter of the elipse.
}
