% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcd-plot.R
\name{lcd}
\alias{lcd}
\title{Liquid Crystal Diagram (LCD) plot}
\usage{
lcd(
  angle_b = 0,
  angle_t = 90,
  angle_function = NULL,
  angle_n = 0,
  n_mol_x = 30,
  n_mol_y = 30,
  x_jitter = 0.4,
  y_jitter = 0.4,
  lc_length = 0.05,
  lc_width = 0.016,
  diagram_aspect = 1,
  seed = NULL,
  surface_b = NULL,
  surface_t = NULL,
  lc_shape = "rectangle",
  ellipse_res = 35,
  return_df = FALSE,
  themeing = TRUE,
  show_function = FALSE,
  lc_lwd = 0.2,
  surface_b_lwd = 0.2,
  surface_t_lwd = 0.2,
  bg_lwd = 0.2,
  lc_fill = "grey80",
  lc_col = "black",
  bg_fill = "white",
  bg_col = "black",
  surface_b_fill = "black",
  surface_b_col = NA,
  surface_t_fill = "black",
  surface_t_col = NA
)
}
\arguments{
\item{angle_b}{Angle of molecules at bottom of plot (degrees, measured anti-clockwise from +ve x)}

\item{angle_t}{Angle of molecules at top of plot (degrees, measured anti-clockwise from +ve x)}

\item{angle_function}{A function that takes a value of the image height (from 0 to 1) and returns the angle of
a molecule at that height}

\item{angle_n}{Noise to be added to angles (a random angle from -angle_n:angle_n is added to each molecule)}

\item{n_mol_x}{Number of molecules in x (default is 30)}

\item{n_mol_y}{Number of molecules in y (default is 30)}

\item{x_jitter}{Jitter to apply to molecule centres in x. Value is applied as a proportion of lc_length (default = 0.4)}

\item{y_jitter}{Jitter to apply to molecule centres in y. Value is applied as a proportion of lc_length (default = 0.4)}

\item{lc_length}{Length of LC molecule (vector of length 1 or n_mol)}

\item{lc_width}{Width of LC molecule (vector of length 1 or n_mol)}

\item{diagram_aspect}{Aspect ratio of diagram (length / width)}

\item{seed}{Random seed for reproducability}

\item{surface_b}{Height of bottom surface layer}

\item{surface_t}{Height of top surface layer}

\item{lc_shape}{One of "rectangle" (default) or "ellipse"}

\item{ellipse_res}{Resolution of ellipse polygons}

\item{return_df}{Should just the dataframe be returned (Default: FALSE)}

\item{themeing}{Should themeing be applied (Default: TRUE)}

\item{show_function}{Logical. Print functional form of angle_function over diagram}

\item{lc_lwd}{Linewidth of LC molecules}

\item{surface_b_lwd}{Linewidth of bottom surface}

\item{surface_t_lwd}{Linewidth of top surface}

\item{bg_lwd}{Linewidth of plot border}

\item{lc_fill}{Fill colour for LC molecules}

\item{lc_col}{Border colour for LC molecules}

\item{bg_fill}{Fill colour for diagram background}

\item{bg_col}{Line colour for diagram background}

\item{surface_b_fill}{Fill colour of bottom surface}

\item{surface_b_col}{Line colour of bottom surface}

\item{surface_t_fill}{Fill colour of top surface}

\item{surface_t_col}{Line colour of top surface}
}
\value{

}
\description{
LCD is a wrapper for gglcd::geom_lc() gglcd::and stat_lc(). It produces simple LC alignmnet diagrams using ggplot2
}
\examples{
lcd(0, 360)
lcd(0, 360, return_df = TRUE)
}
