% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{rects}
\alias{rects}
\title{Rectangles}
\usage{
rects(x, y, angle, length, width)
}
\arguments{
\item{x}{Centre of rectangle in x}

\item{y}{Centre of rectangle in y}

\item{angle}{Angle (degrees) of rotation (measured from positive x to the rectangle long axis)}

\item{length}{Length of rectangle}

\item{width}{Width of rectangle}
}
\value{
A dataframe
}
\description{
Compute the 4 corner coordinates of a rectangle parameterised by its centre position (x, y),
its length, width and angle of rotation
}
