% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcd-plot.R
\name{lcd}
\alias{lcd}
\title{Liquid Crystal Diagram (LCD) plot}
\usage{
lcd(
  theta_b = 0,
  theta_t = 90,
  n_mol = 1000,
  theta_n = 0,
  lc_length = 0.05,
  lc_width = 0.05/3.5,
  diagram_aspect = 1,
  seed = NULL,
  surface_b = NULL,
  surface_t = NULL,
  lc_shape = "rectangle",
  ellipse_res = 35,
  return_df = FALSE,
  themeing = TRUE,
  lc_lwd = 0.2,
  surface_b_lwd = 0.2,
  surface_t_lwd = 0.2,
  bg_lwd = 0.2,
  lc_fill = "grey80",
  lc_col = "black",
  bg_fill = "white",
  bg_col = "black",
  surface_b_fill = "black",
  surface_b_col = NA,
  surface_t_fill = "black",
  surface_t_col = NA
)
}
\arguments{
\item{theta_b}{Theta (degrees) at bottom of diagram}

\item{theta_t}{Theta (degrees) at top of diagram}

\item{n_mol}{Number of molecules for diagram}

\item{theta_n}{Noise angle (degrees). Random noise from -theta_n : theta_n is added to each molecule}

\item{lc_length}{Length of LC molecule (vector of length 1 or n_mol)}

\item{lc_width}{Width of LC molecule (vector of length 1 or n_mol)}

\item{diagram_aspect}{Aspect ratio of diagram (length / width)}

\item{seed}{Random seed for reproducability}

\item{surface_b}{Height of bottom surface layer}

\item{surface_t}{Height of top surface layer}

\item{lc_shape}{One of "rectangle" (default) or "ellipse"}

\item{ellipse_res}{Resolution of ellipse polygons}

\item{return_df}{Should just the dataframe be returned (Default: FALSE)}

\item{themeing}{Should themeing be applied (Default: TRUE)}

\item{lc_lwd}{Linewidth of LC molecules}

\item{surface_b_lwd}{Linewidth of bottom surface}

\item{surface_t_lwd}{Linewidth of top surface}

\item{bg_lwd}{Linewidth of plot border}

\item{lc_fill}{Fill colour for LC molecules}

\item{lc_col}{Border colour for LC molecules}

\item{bg_fill}{Fill colour for diagram background}

\item{bg_col}{Line colour for diagram background}

\item{surface_b_fill}{Fill colour of bottom surface}

\item{surface_b_col}{Line colour of bottom surface}

\item{surface_t_fill}{Fill colour of top surface}

\item{surface_t_col}{Line colour of top surface}
}
\value{

}
\description{
LCD is a wrapper for gglcd::geom_lc() gglcd::and stat_lc(). It produces simple LC alignmnet diagrams
where the molecule angle varies linearly
}
\examples{
lcd(0, 360)
lcd(0, 360, return_df = TRUE)
}
